Refactor: moved `sortedWith` and related free matchers from 'CollectionMatchers.kt' to 'collections/matchers.kt'.
