+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/collections/matchers.kt
@@ -8,7 +8,6 @@ import io.kotlintest.matchers.containAll
 import io.kotlintest.matchers.containsInOrder
 import io.kotlintest.matchers.haveSize
 import io.kotlintest.matchers.singleElement
-import io.kotlintest.matchers.beSortedWith
 import io.kotlintest.neverNullMatcher
 import io.kotlintest.should
 import io.kotlintest.shouldHave
@@ -142,6 +141,26 @@ fun <T> containDuplicates() = object : Matcher<Collection<T>> {
       "Collection should contain duplicates",
       "Collection should not contain duplicates"
   )
+}
+
+
+fun <T : Comparable<T>> beSortedWith(comparator: Comparator<in T>): Matcher<List<T>> = sortedWith(comparator)
+fun <T : Comparable<T>> beSortedWith(cmp: (T, T) -> Int): Matcher<List<T>> = sortedWith(cmp)
+fun <T : Comparable<T>> sortedWith(comparator: Comparator<in T>): Matcher<List<T>> = sortedWith { a, b -> comparator.compare(a, b) }
+fun <T : Comparable<T>> sortedWith(cmp: (T, T) -> Int): Matcher<List<T>> = object : Matcher<List<T>> {
+  override fun test(value: List<T>): Result {
+    val failure = value.withIndex().firstOrNull { (i, it) -> i != value.lastIndex && cmp(it, value[i+1]) > 0 }
+    val snippet = value.joinToString(",", limit = 10)
+    val elementMessage = when (failure) {
+        null -> ""
+        else -> ". Element ${failure.value} at index ${failure.index} shouldn't precede element ${value[failure.index+1]}"
+    }
+    return Result(
+        failure == null,
+        "List [$snippet] should be sorted$elementMessage",
+        "List [$snippet] should not be sorted"
+    )
+  }
 }
 
 fun <T : Comparable<T>> List<T>.shouldBeSorted() = this should beSorted<T>()
